Fixed showing of backup down notif
